Anarix/requirements.txt
fastapi
uvicorn
pydantic
pandas
matplotlib
seaborn
plotly
python-dotenv
httpx
streamlit 